# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This repository is an example of an API made in Go with the To Do List theme.",
        "title": "To Do List API",
        "contact": {
            "name": "Lucas Santos",
            "url": "https://github.com/devlucassantos"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
...
        "/user/{userId}/collection": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Route that allows searching all user collections in the system",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collection"
                ],
                "summary": "Lists all user collections",
                "operationId": "FindAllCollections",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful request",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.SwaggerCollectionResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "The user has made a bad request",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerValidationErrorResponse"
                        }
                    },
                    "401": {
                        "description": "The user is not authorized to make this request",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerUnauthorizedResponse"
                        }
                    },
                    "403": {
                        "description": "The user does not have access to this information",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerForbiddenResponse"
                        }
                    },
                    "404": {
                        "description": "The user has requested a non-existent resource",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerNotFoundErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Some entered data could not be processed because it is not valid",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "An unexpected server error has occurred",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerGenericErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Route that allows registering a collection in the system. To register a collection it is necessary to inform the following data in the body of the request:\n|   Name   |  Type  |   Required  | Description      |\n|----------|--------|-------------|------------------|\n|   name   | string |      x      | Collection name |",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collection"
                ],
                "summary": "Create a collection",
                "operationId": "CreateCollection",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "JSON responsible for sending all collection registration data to the database",
                        "name": "authJson",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SwaggerCollectionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Collection successfully registered",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerIdResponse"
                        }
                    },
                    "400": {
                        "description": "The user has made a bad request",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerValidationErrorResponse"
                        }
                    },
                    "401": {
                        "description": "The user is not authorized to make this request",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerUnauthorizedResponse"
                        }
                    },
                    "403": {
                        "description": "The user does not have access to this information",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerForbiddenResponse"
                        }
                    },
                    "422": {
                        "description": "Some entered data could not be processed because it is not valid",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "An unexpected server error has occurred",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerGenericErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/{userId}/collection/{collectionId}": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Route that allows searching a collection registered in the system by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collection"
                ],
                "summary": "Search a collection's data by ID",
                "operationId": "FindByCollectionId",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Collection ID",
                        "name": "collectionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful request",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerCollectionResponse"
                        }
                    },
                    "400": {
                        "description": "The user has made a bad request",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerValidationErrorResponse"
                        }
                    },
                    "401": {
                        "description": "The user is not authorized to make this request",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerUnauthorizedResponse"
                        }
                    },
                    "403": {
                        "description": "The user does not have access to this information",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerForbiddenResponse"
                        }
                    },
                    "404": {
                        "description": "The user has requested a non-existent resource",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerNotFoundErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Some entered data could not be processed because it is not valid",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "An unexpected server error has occurred",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerGenericErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Route that allows editing a collection in the system. To edit a collection it is necessary to inform the following data:\n|   Name   |  Type  |   Required  | Description\t     |\n|----------|--------|-------------|------------------|\n|   name   | string |      x      | Collection name  |",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collection"
                ],
                "summary": "Update a collection",
                "operationId": "UpdateCollection",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Collection ID",
                        "name": "collectionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "JSON responsible for sending the data needed to update the collection in the database",
                        "name": "authJson",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SwaggerCollectionRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Collection successfully edited"
                    },
                    "400": {
                        "description": "The user has made a bad request",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerValidationErrorResponse"
                        }
                    },
                    "401": {
                        "description": "The user is not authorized to make this request",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerUnauthorizedResponse"
                        }
                    },
                    "403": {
                        "description": "The user does not have access to this information",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerForbiddenResponse"
                        }
                    },
                    "404": {
                        "description": "The user has requested a non-existent resource",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerNotFoundErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Some entered data could not be processed because it is not valid",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "An unexpected server error has occurred",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerGenericErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Route that allows deleting a collection registered in the system",
                "tags": [
                    "Collection"
                ],
                "summary": "Delete a collection",
                "operationId": "DeleteCollection",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Collection ID",
                        "name": "collectionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Collection successfully deleted"
                    },
                    "400": {
                        "description": "The user has made a bad request",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerValidationErrorResponse"
                        }
                    },
                    "401": {
                        "description": "The user is not authorized to make this request",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerUnauthorizedResponse"
                        }
                    },
                    "403": {
                        "description": "The user does not have access to this information",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerForbiddenResponse"
                        }
                    },
                    "404": {
                        "description": "The user has requested a non-existent resource",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerNotFoundErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Some entered data could not be processed because it is not valid",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "An unexpected server error has occurred",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerGenericErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "request.SwaggerCollectionRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Collection example"
                }
            }
        },
...
        "response.SwaggerCollectionResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Collection example"
                }
            }
        },
...
        "response.SwaggerForbiddenResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Oops! You do not have access to this information."
                }
            }
        },
        "response.SwaggerGenericErrorResponse": {
            "type": "object",
            "properties": {
                "error_msg": {
                    "type": "string",
                    "example": "Oops! An unexpected error has occurred."
                }
            }
        },
        "response.SwaggerIdResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "response.SwaggerInvalidField": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Description example"
                },
                "name": {
                    "type": "string",
                    "example": "Field example"
                }
            }
        },
        "response.SwaggerNotFoundErrorResponse": {
            "type": "object",
            "properties": {
                "error_msg": {
                    "type": "string",
                    "example": "Not Found"
                }
            }
        },
...
        "response.SwaggerUnauthorizedResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Oops! You are not authorized."
                }
            }
        },
        "response.SwaggerValidationErrorResponse": {
            "type": "object",
            "properties": {
                "error_msg": {
                    "type": "string",
                    "example": "Some of the data entered is invalid."
                },
                "invalid_fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.SwaggerInvalidField"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "bearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}